refactor: restructure app by moving providers to a dedicated directory and removing unused files
