+++ src/app/providers/mantine.tsx
@@ -1,0 +1,8 @@
+import { MantineProvider } from '@mantine/core';
+import type { PropsWithChildren } from 'react';
+
+import { theme } from '@/styles/mantine.theme';
+
+export function AppMantineProvider({ children }: Readonly<PropsWithChildren>) {
+  return <MantineProvider theme={theme}>{children}</MantineProvider>;
+}
